Rename files in metadata module
