+++ streamrip/metadata/__init__.py
@@ -1,10 +1,10 @@
 """Manages the information that will be embeded in the audio file."""
 from . import util
-from .album_metadata import AlbumMetadata
-from .artist_metadata import ArtistMetadata
+from .album import AlbumInfo, AlbumMetadata
+from .artist import ArtistMetadata
 from .covers import Covers
-from .label_metadata import LabelMetadata
-from .playlist_metadata import PlaylistMetadata
+from .label import LabelMetadata
+from .playlist import PlaylistMetadata
 from .search_results import (
     AlbumSummary,
     ArtistSummary,
@@ -15,11 +15,13 @@ from .search_results import (
     TrackSummary,
 )
 from .tagger import tag_file
-from .track_metadata import TrackMetadata
+from .track import TrackInfo, TrackMetadata
 
 __all__ = [
     "AlbumMetadata",
     "ArtistMetadata",
+    "AlbumInfo",
+    "TrackInfo",
     "LabelMetadata",
     "TrackMetadata",
     "PlaylistMetadata",
