+++ streamrip/metadata/album.py
