+++ streamrip/metadata/artist.py
